+++ src/wb_deferred.ts
@@ -1,6 +1,6 @@
 /*!
- * Wunderbaum - deferred
- * Copyright (c) 2021-2025, Martin Wendt. Released under the MIT license.
+ * Wunderbaum - wb_deferred
+ * Copyright (c) 2021-2026, Martin Wendt. Released under the MIT license.
  * @VERSION, @DATE (https://github.com/mar10/wunderbaum)
  */