+++ src/wb_ext_dnd.ts
@@ -1,10 +1,10 @@
 /*!
  * Wunderbaum - ext-dnd
- * Copyright (c) 2021-2025, Martin Wendt. Released under the MIT license.
+ * Copyright (c) 2021-2026, Martin Wendt. Released under the MIT license.
  * @VERSION, @DATE (https://github.com/mar10/wunderbaum)
  */
-import * as util from "./util";
-import { EventCallbackType, onEvent } from "./util";
+import * as util from "./wb_util";
+import { EventCallbackType, onEvent } from "./wb_util";
 import { Wunderbaum } from "./wunderbaum";
 import { WunderbaumExtension } from "./wb_extension_base";
 import { WunderbaumNode } from "./wb_node";
@@ -13,8 +13,8 @@ import {
   DropEffectType,
   DropRegionType,
   DropRegionTypeSet,
-} from "./types";
-import { DebouncedFunction, throttle } from "./debounce";
+} from "./wb_types";
+import { DebouncedFunction, throttle } from "./wb_debounce";
 
 const nodeMimeType = "application/x-wunderbaum-node";