+++ src/wb_options.ts
@@ -1,6 +1,6 @@
 /*!
  * Wunderbaum - options
- * Copyright (c) 2021-2025, Martin Wendt. Released under the MIT license.
+ * Copyright (c) 2021-2026, Martin Wendt. Released under the MIT license.
  * @VERSION, @DATE (https://github.com/mar10/wunderbaum)
  */
 
@@ -39,7 +39,7 @@ import {
   WbRenderEventType,
   WbSelectEventType,
   WbTreeEventType,
-} from "./types";
+} from "./wb_types";
 
 /**
  * Properties of {@link wunderbaum.Wunderbaum.options}.